Исправление ошибок. Добавление метода пересчета общего времени.

+++ Worktime.Core/CRUD/LineProcessor.cs
@@ -1,4 +1,5 @@
 ﻿using Microsoft.EntityFrameworkCore;
+using System.Threading.Tasks;
 using Worktime.Core.Models;
 
 namespace Worktime.Core.CRUD
@@ -20,9 +21,7 @@ namespace Worktime.Core.CRUD
 
             if (line.BeginTime > line.EndTime)
                 line.EndTime = line.BeginTime;
-            double time = (line.EndTime - line.BeginTime).Minutes;
-            if (time > 0)
-                time /= 60;
+            double time = Math.Round((line.EndTime - line.BeginTime).TotalHours, 2);
             line.Time = time;
 
             result.Items.Add(_db.Lines.Add(line).Entity);
@@ -34,7 +33,10 @@ namespace Worktime.Core.CRUD
             Result<WTLine> result = new();
             result = CreateOne(line, result);
             if (result.Success)
+            {
                 _db.SaveChanges();
+                RecalculateTotalTime(result.Items.First());
+            }
             return result;
         }
         public Result<WTLine> Create(IEnumerable<WTLine> lines)
@@ -42,7 +44,10 @@ namespace Worktime.Core.CRUD
             Result<WTLine> result = new();
             result = ResultGeneric.Execute(lines, result, CreateOne);
             if (result.Success && lines.Any())
+            {
                 _db.SaveChanges();
+                RecalculateTotalTime(result.Items.First());
+            }
             return result;
         }
         #endregion
@@ -70,13 +75,10 @@ namespace Worktime.Core.CRUD
             line.EndTime = newLine.EndTime;
             if (line.BeginTime > line.EndTime)
                 line.EndTime = line.BeginTime;
-            double time = (line.EndTime - line.BeginTime).Minutes;
-            if (time > 0)
-                time /= 60;
+            double time = Math.Round((line.EndTime - line.BeginTime).TotalHours, 2);
             line.Time = time;
             line.WTTaskId = newLine.WTTaskId;
             line.Task = task;
-            task.TotalTime += time;
 
             _db.Tasks.Update(task);
             result.Items.Add(_db.Lines.Update(line).Entity);
@@ -88,7 +90,10 @@ namespace Worktime.Core.CRUD
             Result<WTLine> result = new();
             result = UpdateOne(newLine, result);
             if (result.Success)
+            {
                 _db.SaveChanges();
+                RecalculateTotalTime(result.Items.First());
+            }
             return result;
         }
         public Result<WTLine> Update(IEnumerable<WTLine> lines)
@@ -96,7 +101,10 @@ namespace Worktime.Core.CRUD
             Result<WTLine> result = new();
             result = ResultGeneric.Execute(lines, result, UpdateOne);
             if (result.Success && lines.Any())
+            {
                 _db.SaveChanges();
+                RecalculateTotalTime(result.Items.First());
+            }
             return result;
         }
         #endregion
@@ -104,9 +112,12 @@ namespace Worktime.Core.CRUD
         #region Delete
         private Result<WTLine> DeleteOne(WTLine item, Result<WTLine> result)
         {
+            var task = _db.Tasks.Find(item.WTTaskId);
             var line = _db.Lines.Find(item.Id);
             if (line == null)
                 return new Result<WTLine>() { Success = false, Message = "Line was not found!" };
+            if (task == null)
+                return new Result<WTLine>() { Success = false, Message = "Task was not found!" };
 
             result.Items.Add(_db.Lines.Remove(line).Entity);
             result.Success = true;
@@ -117,7 +128,10 @@ namespace Worktime.Core.CRUD
             Result<WTLine> result = new();
             result = DeleteOne(item, result);
             if (result.Success)
+            {
                 _db.SaveChanges();
+                RecalculateTotalTime(result.Items.First());
+            }
             return result;
         }
         public Result<WTLine> Delete(IEnumerable<WTLine> items)
@@ -125,9 +139,23 @@ namespace Worktime.Core.CRUD
             Result<WTLine> result = new();
             result = ResultGeneric.Execute(items, result, DeleteOne);
             if (result.Success && items.Any())
+            {
                 _db.SaveChanges();
+                RecalculateTotalTime(result.Items.First());
+            }
             return result;
-        } 
+        }
         #endregion
+
+        private void RecalculateTotalTime(WTLine item)
+        {
+            var task = _db.Tasks.Find(item.WTTaskId);
+            if(task != null)
+            {
+                task.TotalTime = _db.Lines.Where(x => x.WTTaskId == task.Id).Sum(x => x.Time);
+                _db.Tasks.Update(task);
+                _db.SaveChanges();
+            }
+        }
     }
 }
